@@ -1,9 +1,9 @@
 package com.hellohasan.sqlite_multiple_three_tables_crud.features.subject_crud.subject_create;
 
 import android.app.Dialog;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -37,7 +37,7 @@ public class SubjectCreateDialogFragment extends DialogFragment {
         args.putString("title", title);
         subjectCreateDialogFragment.setArguments(args);
 
-        subjectCreateDialogFragment.setStyle(android.support.v4.app.DialogFragment.STYLE_NORMAL, R.style.CustomDialog);
+        subjectCreateDialogFragment.setStyle(DialogFragment.STYLE_NORMAL, R.style.CustomDialog);
 
         return subjectCreateDialogFragment;
     }
