@@ -2,7 +2,7 @@ package com.hellohasan.sqlite_multiple_three_tables_crud.features.subject_crud.s
 
 import android.app.Dialog;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
